requirements.txt
pillow==11.0.0
pandas==2.2.3
tensorflow==2.17.1
pytest-dotenv==0.5.2
scikit-learn==1.6.0
matplotlib==3.10.0